#3 遅延表示を実装
